Revert "add Telegram notification"
This reverts commit 178ae8b0c610fb98531deab1787b6046c69ffe44.
