+++ README.md
@@ -1,11 +1,10 @@
 # Создание триггера для бюджетов, который вызывает функцию Cloud Functions   
 
-Репозиторий создан для реализации сценария остановки виртуальных машин и отправки уведомления в Telegram в следствии вызова триггера для бюджетов.
+Репозиторий создан для реализации сценария остановки виртуальных машин в следствии вызова триггера для бюджетов.
 
 Инструменты, с которыми вы будете работать в рамках этого репозитория:
 * CLI Yandex Cloud;
 * Yandex Cloud Functions;
 * Yandex Compute;
-* Yandex Message Queue;
 
-В папке `steps` содержится последовательность шагов необходимых для настройки решения с соответствующими пояснениями, в конце некоторых шагов есть ссылка на сопутствующее видео.
+В папке `steps` содержится последовательность шагов необходимых для настройки решения с соответствующими пояснениями, в конце каждого шага есть ссылка на сопутствующее видео.
