+++ steps/3-create-budgets/README.md
@@ -1,4 +1,4 @@
-# Шаг 2. Создание бюджета
+# Шаг 3. Создать бюджет
 
 Для создания бюджета необходима роль `editor` у пользователя. Для получения оповещений достаточно роли `viewer`.
