+++ steps/5-create-compute-instance/README.md
@@ -1,4 +1,4 @@
-# Шаг 9. Создание Serverless Container из Docker-образа
+# Шаг 5. Создание Serverless Container из Docker-образа
 
 Создадим несколько виртуальных машин, чтобы имитировать потребление в облаке. Для этого нам потребуется уточнить значение сетей и подсетей.
