+++ steps/6-deleting-resources/README.md
@@ -1,4 +1,4 @@
-# Шаг 10. Удаление ресурсов
+# Шаг 6. Удаление ресурсов
 
 По завершению работы удалите все используемые вами ресурсы.
