--- steps/6-register-telegram-bot/README.md
@@ -1,15 +1,0 @@
-# Шаг 6. Регистрация телеграм бота
-
-Перейдите в [официального бота телеграм](https://t.me/BotFather) для регистрации своего бота. Введите команду /newbot, далее, следуя инструкциям, введите желаемое название бота, а также его username. В результате вы получите HTTP API token, его запишите в переменную окружения:
-
-    echo "export TELEGRAM_BOT_API_TOKEN=<api_token>" >> ~/.bashrc && . ~/.bashrc 
-
-Чтобы бот мог отправить пользователю сообещние он должен знать его ChatID, а для его получения нужно чтобы пользователь начал с ботом диалог. В данном примере мы передадим ChatID единственного пользователя боту в переменных окружения. Для этого из текущей папки 6-register-telegram-bot выполните команду:
-
-    go run .
-
-Она запустит бота, который на все сообщения будет возвращать ChatID текущего пользователя. Откройте созданного бота в телеграм, отправьте ему сообщение и запишите полученный ChatID:
-
-    echo "export TELEGRAM_BOT_CHAT_ID=<chat_id>" >> ~/.bashrc && . ~/.bashrc 
-
-В конце завершите работу бота нажав комбинацию Ctrl+C.
