--- steps/7-create-budget-queue-handler/go.mod
@@ -1,7 +1,0 @@
-module budget_alert
-
-go 1.19
-
-require github.com/go-telegram-bot-api/telegram-bot-api/v5 v5.5.1
-
-require github.com/pkg/errors v0.9.1 // indirect
